requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
tensorflow==2.10.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.23.5
pickle-mixin==1.0.2
pytest
httpx
requests==2.26.0
pymongo[srv]
python-dotenv
webdriver-manager==3.4.2
selenium==4.1.0